% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datautils.r
\name{augment_data}
\alias{augment_data}
\title{Augment Data}
\usage{
augment_data(df, wy_categories = NULL, event_years = NULL)
}
\arguments{
\item{df}{A dataset, i.e., output of \code{\link[=read_wat_binary]{read_wat_binary()}}.}

\item{wy_categories}{A data frame of categories. At minimum, the fields
"wateryear" and "wycategory" are required. Additional fields
"alt", "run" and "event" are also used to assign categories to
\code{df} if they are present in \code{categories}.}

\item{event_years}{A dataframe containing at minimum the fields
"event" and "baseyear". Fields "alt", "run", and
"wateryear" will also be used to attach baseyear information if
available.}
}
\value{
The standardized, augmented dataset.
}
\description{
Augment a dataset by adding water year and reference date,
and attaching supplemental information such as water year
categories, and base years.
}
