% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileioutils.r
\name{bin_format}
\alias{bin_format}
\title{Parse Binary Format String}
\usage{
bin_format(format)
}
\arguments{
\item{format}{The format string.}
}
\value{
A list of format components.
}
\description{
Parse the binary file format string for use with \code{\link[=readBin]{readBin()}}.
}
\keyword{internal}
