% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileioutils.r
\name{binary_name_to_path}
\alias{binary_name_to_path}
\title{Binary Name to Path}
\usage{
binary_name_to_path(file, binsep = "__")
}
\arguments{
\item{file}{The path to the binary file.}

\item{binsep}{The separator used in WAT binary files to replace the
path separator "/".}
}
\value{
The equivalent path.
}
\description{
Convert a binary file name to a path.
}
\keyword{internal}
