% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileioutils.r
\name{get_file_extension}
\alias{get_file_extension}
\alias{strip_file_extension}
\title{File Extension Handling}
\usage{
get_file_extension(file)

strip_file_extension(file)
}
\arguments{
\item{file}{A file name or path.}
}
\description{
Extract or strip file extensions.
}
\keyword{internal}
