% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileioutils.r
\name{path_to_binary_name}
\alias{path_to_binary_name}
\title{Path To Binary File Name}
\usage{
path_to_binary_name(path, fileext = ".dat", binsep = "__")
}
\arguments{
\item{path}{A vector of paths.}

\item{fileext}{a non-empty character vector giving the file extension}

\item{binsep}{The separator used in WAT binary files to replace the
path separator "/".}
}
\value{
A filename
}
\description{
Convert a path to its standard binary file name.
}
\keyword{internal}
