% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datautils.r
\name{prepare_data}
\alias{prepare_data}
\title{Prepare Data}
\usage{
prepare_data(df, time_window = NULL, wy_category = NULL, base_year = NULL)
}
\arguments{
\item{df}{A dataset, i.e., output of
\code{\link[=read_wat_binary]{read_wat_binary()}}.}

\item{time_window}{A vector of time windows for use with
\code{\link[watplotter.core:extract_window]{watplotter.core::extract_window()}}.}

\item{wy_category}{A vector of water year categories for use with
\code{\link[watplotter.core:extract_wycategory]{watplotter.core::extract_wycategory()}}. Ignored if \code{NA} or \code{NULL}.}

\item{base_year}{A vector of base years for use with
\code{\link[watplotter.core:extract_baseyear]{watplotter.core::extract_baseyear()}}. Ignored if \code{NA} or \code{NULL}.}
}
\value{
The filtered, standardized dataset.
}
\description{
Prepare data for analysis by adding water year, reference date,
and water year categories, and filtering by time window and
base year(s).
}
