% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileioutils.r
\name{read_bin_file}
\alias{read_bin_file}
\title{Read Binary File}
\usage{
read_bin_file(file, tz = "UTC", format = ">i4;>i4;>f4")
}
\arguments{
\item{file}{The path to the binary file.}

\item{tz}{The timezone to use when parsing datetimes.}

\item{format}{The format string, in form \code{"(endian)(type)(bytes)"}.
fields are separated with \code{";"}}
}
\value{
A tibble with fields "date", "event", "value".
}
\description{
Read a binary file into a dataframe.
}
\details{
The input binary file is expected to match this byte order:
\itemize{
\item bytes 1-4: integer (date)
\item bytes 5-8: integer (event)
\item bytes 9-12: float (value)
}
}
\keyword{internal}
