% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileioutils.r
\name{read_wat_binary}
\alias{read_wat_binary}
\title{Read WAT Binary File}
\usage{
read_wat_binary(
  file,
  ...,
  tz = "UTC",
  format = ">i4;>i4;>f4",
  skip_missing = FALSE
)
}
\arguments{
\item{file}{A vector of file paths, or a list of raw vectors.}

\item{...}{Named identifier columns to be
joined to the return data, e.g., the alternative name or other
information. Each argument must match length and order of \code{file}.}

\item{tz}{The timezone to use when parsing datetimes.}

\item{format}{The format string, in form \code{"(endian)(type)(bytes)"}.
fields are separated with \code{";"}}

\item{skip_missing}{If \code{TRUE}, skip over missing files with a
warning.}
}
\value{
A tibble with fields "path", "date", "event", "value" and
(if supplied) additional fields from \code{label}.
}
\description{
Read a WAT binary file into a dataframe.
}
