% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\docType{package}
\name{watplotter.io-package}
\alias{watplotter.io}
\alias{watplotter.io-package}
\title{watplotter.io: WAT Plotter Input/Output}
\description{
Read and write WAT binary files and manipulate DSS pathnames.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://nwd-wmshare.nwd.usace.army.mil:3000/git/DataMgmt/watplotter.io}
  \item Report bugs at \url{http://nwd-wmshare.nwd.usace.army.mil:3000/git/DataMgmt/watplotter.io/issues}
}

}
\author{
\strong{Maintainer}: Michael Koohafkan \email{Michael.C.Koohafkan@usace.army.mil}

}
\keyword{internal}
