% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileioutils.r
\name{write_bin_file}
\alias{write_bin_file}
\title{Write Binary File}
\usage{
write_bin_file(df, file)
}
\arguments{
\item{df}{A dataframe. Must include fields "date", "event",
and "value".}

\item{file}{The file path to write to.}
}
\value{
file, invisibly.
}
\description{
Write a dataframe to a WAT binary file.
}
\details{
The output binary file is structured as 12-byte sequences:
\itemize{
\item bytes 1-4: integer (date)
\item bytes 5-8: integer (event)
\item bytes 9-12: float (value)
}
}
\keyword{internal}
