% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileioutils.r
\name{write_wat_binary}
\alias{write_wat_binary}
\title{Write a dataframe to one or binary files, as determined by the
dataframe's \code{"path"} field.}
\usage{
write_wat_binary(df, output_dir = ".", validate_paths = FALSE)
}
\arguments{
\item{df}{A dataframe. Must include field "path".}

\item{output_dir}{The directory to write the binary files to.}

\item{validate_paths}{If \code{TRUE}, validate the format of all paths
before writing.}
}
\value{
(invisibly) A vector of binary file paths.
}
\description{
Write a dataframe to one or binary files, as determined by the
dataframe's \code{"path"} field.
}
